+++ src/proxy-ui-api/frontend/src/components/layout/TabsBase.vue
