+++ src/proxy-ui-api/frontend/src/components/ui/SnackbarMixin.vue
