+++ src/proxy-ui-api/frontend/src/components/ui/WarningDialog.vue
@@ -57,6 +57,6 @@ export default Vue.extend({
 </script>
 
 <style lang="scss" scoped>
-@import '../assets/dialogs';
+@import '../../assets/dialogs';
 </style>
