+++ src/proxy-ui-api/frontend/src/router.ts
@@ -1,7 +1,8 @@
 import Router, { Route } from 'vue-router';
 import { sync } from 'vuex-router-sync';
-import AppLogin from './views/AppLogin.vue';
-import Base from './views/Base.vue';
+import TabsBase from '@/components/layout/TabsBase.vue';
+import AppLogin from '@/views/AppLogin.vue';
+import AppBase from '@/views/AppBase.vue';
 import Clients from '@/views/Clients/Clients.vue';
 import Client from '@/views/Clients/Client.vue';
 import KeysAndCertificates from '@/views/KeysAndCertificates/KeysAndCertificates.vue';
@@ -12,24 +13,23 @@ import Settings from '@/views/Settings/Settings.vue';
 import SystemParameters from '@/views/Settings/SystemParameters.vue';
 import BackupAndRestore from '@/views/Settings/BackupAndRestore.vue';
 import Diagnostics from '@/views/Diagnostics/Diagnostics.vue';
-import AddSubsystem from './views/AddSubsystem.vue';
-import AddClient from './views/AddClient.vue';
-import Subsystem from './views/Clients/Subsystem.vue';
+import AddSubsystem from '@/views/AddSubsystem.vue';
+import AddClient from '@/views/AddClient.vue';
+import Subsystem from '@/views/Clients/Subsystem.vue';
 import ClientDetails from '@/views/Clients/Details/ClientDetails.vue';
 import InternalServers from '@/views/Clients/InternalServers/InternalServers.vue';
 import Services from '@/views/Clients/Services/Services.vue';
 import ServiceClients from '@/views/Clients/ServiceClients/ServiceClients.vue';
 import LocalGroups from '@/views/Clients/LocalGroups/LocalGroups.vue';
-import TabsBase from '@/views/TabsBase.vue';
 import ClientTlsCertificate from '@/views/ClientTlsCertificate.vue';
-import Error from '@/views/Error.vue';
+import AppError from '@/views/AppError.vue';
 import LocalGroup from '@/views/LocalGroup/LocalGroup.vue';
 import ServiceDescriptionDetails from '@/views/ServiceDescriptionDetails.vue';
 import TokenDetails from '@/views/TokenDetails.vue';
 import KeyDetails from '@/views/KeyDetails.vue';
 import CertificateDetails from '@/views/CertificateDetails.vue';
 import Service from '@/views/Service/Service.vue';
-import store from './store';
+import store from '@/store';
 import { RouteName, Permissions } from '@/global';
 
 
@@ -37,7 +37,7 @@ const router = new Router({
   routes: [
     {
       path: '/',
-      component: Base,
+      component: AppBase,
       children: [
         {
           path: '/keys',
@@ -266,7 +266,7 @@ const router = new Router({
     },
     {
       path: '*',
-      component: Error,
+      component: AppError,
     },
   ],
 });
