+++ src/proxy-ui-api/frontend/src/views/AppBase.vue
