+++ src/proxy-ui-api/frontend/src/views/AppError.vue
