+++ src/proxy-ui-api/frontend/src/views/Clients/Services/Services.vue
@@ -140,7 +140,7 @@ import Expandable from '@/components/ui/Expandable.vue';
 import AddWsdlDialog from './AddWsdlDialog.vue';
 import AddRestDialog from './AddRestDialog.vue';
 import DisableServiceDescDialog from './DisableServiceDescDialog.vue';
-import WarningDialog from '@/components/WarningDialog.vue';
+import WarningDialog from '@/components/ui/WarningDialog.vue';
 import ServiceIcon from '@/components/ui/ServiceIcon.vue';
 
 import _ from 'lodash';
