+++ src/proxy-ui-api/frontend/src/views/ServiceDescriptionDetails.vue
@@ -111,7 +111,7 @@ import { Permissions } from '@/global';
 import * as api from '@/util/api';
 import SubViewTitle from '@/components/ui/SubViewTitle.vue';
 import ConfirmDialog from '@/components/ui/ConfirmDialog.vue';
-import WarningDialog from '@/components/WarningDialog.vue';
+import WarningDialog from '@/components/ui/WarningDialog.vue';
 import LargeButton from '@/components/ui/LargeButton.vue';
 
 export default Vue.extend({
